+++ src/setupTests.ts
@@ -1,0 +1,13 @@
+// jest-dom adds custom jest matchers for asserting on DOM nodes.
+// allows you to do things like:
+// expect(element).toHaveTextContent(/react/i)
+// learn more: https://github.com/testing-library/jest-dom
+import { expect, afterEach } from 'vitest'
+import { cleanup } from '@testing-library/react';
+import matchers from '@testing-library/jest-dom/matchers'
+
+expect.extend(matchers)
+
+afterEach(() => {
+  cleanup();
+});
