+++ vite.config.ts
@@ -1,7 +1,16 @@
-import { defineConfig } from 'vite'
+import { defineConfig } from 'vitest/config'
 import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  test: {
+    environment: 'jsdom', 
+    setupFiles: 'src/setupTests.ts',
+    coverage: {
+      enabled: true,
+      provider: 'istanbul',
+      all: true
+    }
+  }
 })
